feat(provider api): support return fields
