+++ api/package.json
@@ -5,7 +5,9 @@
     "express-session": "^1.16.1",
     "lodash": "^4.17.11",
     "micro": "^9.3.4",
+    "micro-compose": "0.0.3",
     "micro-query": "^0.3.0",
+    "micro-visualize": "^3.0.1",
     "microrouter": "^3.1.3",
     "mongoose": "^5.5.12",
     "passport": "^0.4.0",
